added a sqlite3 database to allow more customizable features

+++ main.py
@@ -3,31 +3,51 @@ import os
 import discord
 from dotenv import load_dotenv
 from utils import is_substring
+from discord.ext import commands
+import sqlite3
 
 load_dotenv()
 TOKEN = os.getenv('DISCORD_TOKEN')
 GUILD = os.getenv('GUILD_TOKEN')
 client = discord.Client()
+bot = commands.Bot(command_prefix='!')
 
 
-@client.event
+@bot.command()
+async def test(ctx, arg):
+    await ctx.send(arg)
+
+
+@bot.event
 async def on_ready():
-    for guild in client.guilds:
+    for guild in bot.guilds:
         if guild.name == GUILD:
             break
-    print(f'{client.user} has connected to Discord!\n'
+    print(f'{bot.user} has connected to Discord!\n'
           f'Discord ID:{guild.id}')
     members = [member.name for member in guild.members]
     print(f'\n{members}')
 
 
-@client.event
+@bot.event
 async def on_message(message):
-    swears = ['zbi', 'idk fih', 'fuck', 'shit']
+    # swears = ['zbi', 'idk fih', 'fuck', 'shit', 'ass']
+    if not message:
+        return
+    db = sqlite3.connect('main.sqlite')
+    cursor = db.cursor()
+    try:
+        guild_id = message.guild.id
+    except AttributeError:
+        return
+    cursor.execute(f'SELECT word FROM badwords WHERE guild_id={guild_id}')
+    swears = [swear[0] for swear in cursor.fetchall()]
+    if not swears:
+        return
     if client.user == message.author:
         return
     for swearword in swears:
-        if is_substring(message.content.lower(), swearword):
+        if not message.channel.is_nsfw() and is_substring(message.content.lower(), swearword):
             await message.delete()
             await message.author.create_dm()
             await message.author.dm_channel.send(
@@ -35,4 +55,4 @@ async def on_message(message):
             )
 
 
-client.run(TOKEN)
+bot.run(TOKEN)
